requirements.txt
Flask==2.3.2
beautifulsoup4==4.12.2
gunicorn==21.2.0
pandas
requests
python-dotenv